fixed lint
fixed coverage

added test cases

fixed lint

fixed test
