+++ packages/config/src/defaults.js
@@ -1,4 +1,4 @@
-import { merge } from './utils/index.js';
+import { merge, sanitizeObject } from './utils/index.js';
 import { getSchema } from './validate.js';
 
 const { isArray } = Array;
@@ -16,38 +16,12 @@ function getDefaultsFromSchema(schema) {
   } else if (schema.type === 'object' && schema.properties) {
     // return an object of default properties
     return entries(schema.properties).reduce((acc, [prop, schema]) => {
-      if(!isSafeKey(prop)){
-        return acc;
-      }
       let def = getDefaultsFromSchema(schema);
       return def != null ? assign(acc || {}, { [prop]: def }) : acc;
     }, undefined);
   } else {
     return undefined;
   }
-}
-
-// Utility function to prevent prototype pollution
-function isSafeKey(key) {
-  const unsafeKeys = ['__proto__', 'constructor', 'prototype', 'toString', 'valueOf',
-                      '__defineGetter__', '__defineSetter__', '__lookupGetter__', '__lookupSetter__'];
-  return !unsafeKeys.includes(key);
-}
-
-
-function sanitizeObject(obj) {
-  if (!obj || typeof obj !== 'object' || isArray(obj)) {
-    return obj; 
-  }
-
-  const sanitized = {};
-  for (const key in obj) {
-    if (isSafeKey(key)) {
-      sanitized[key] = sanitizeObject(obj[key]);
-    }
-  }
-
-  return sanitized;
 }
 
 export function getDefaults(overrides = {}) {
